Refactor: separate cluster creation form from cluster creation page
